Create uncoarsener base class and move function bodies out of headers
