+++ mt-kahypar/partition/coarsening/CMakeLists.txt
@@ -1,6 +1,8 @@
 set(CoarseningSources
         deterministic_multilevel_coarsener.cpp
         uncoarsening.cpp
+        multilevel_uncoarsener.cpp
+        nlevel_uncoarsener.cpp
         )
 
 foreach(modtarget IN LISTS TARGETS_WANTING_ALL_SOURCES)
